Add files via upload
Data was provided as a zip with multiple csv files. Pre-processing of this data first involves adding a numerical category tag to describe the type of coin 1 for bitcoin cash 2 for bitcoin 3 for ethereum 4 for litecoin and 5 for ripple. Next each of these csv's must be combined to form one master list of the coins totaling 1,314,830 points of data. Some columns from the master list need to be trimmed off because they are not relevant to our project. Of these we will eliminate the Order number, buy order number, sell order number, and type.